Move exceptions to separate file
