+++ psutil/_pswindows.py
@@ -45,6 +45,9 @@ from ._compat import lru_cache
 from ._compat import PY3
 from ._compat import unicode
 from ._compat import xrange
+from ._exceptions import NoSuchProcess
+from ._exceptions import AccessDenied
+from ._exceptions import TimeoutExpired
 from ._psutil_windows import ABOVE_NORMAL_PRIORITY_CLASS
 from ._psutil_windows import BELOW_NORMAL_PRIORITY_CLASS
 from ._psutil_windows import HIGH_PRIORITY_CLASS
@@ -138,11 +141,6 @@ pinfo_map = dict(
     peak_pagefile=20,
     mem_private=21,
 )
-
-# these get overwritten on "import psutil" from the __init__.py file
-NoSuchProcess = None
-AccessDenied = None
-TimeoutExpired = None
 
 
 # =====================================================================
